refactor: move from msgpack to protobuf
This commit rewrites most of the database core and TypeScript SDK to
support protobuf over msgpack and drop the custom parser in favor of the
built-in parser generated for the proto file.

The AST is moved to `proto/rulo.proto`, which is used to generate server
and client code as well.

Besides the refactoring, some new features are implemented to allow
better SDK usage. There are some defined and unimplemented features as
well, such as authentication.

The following new features ended up in the refactoring:
- Skip
- Limit
- Match expression (regex)
- Update
- OrderBy

Signed-off-by: Gabor Boros <gabor@opcotech.com>