feat: implement plucking fields from a document (#17)
* feat: implement plucking fields from a document

Implement `pluck` functionality to select and return only a given set of
fields from a document.

Resolves: #9
Signed-off-by: Gabor Boros <gabor@opcotech.com>

* fix: return cursor or object based on context

Return a cursor or an object depending on the type of input data. If
pluck was called on an operation that produces a single object, return
the object, otherwise return a collection of objects.

Signed-off-by: Gabor Boros <gabor@opcotech.com>

* style: resolve clippy errors

Signed-off-by: Gabor Boros <gabor@opcotech.com>

---------

Signed-off-by: Gabor Boros <gabor@opcotech.com>